requirements.txt
aiofiles==23.1.0
email_validator==1.3.1
fastapi==0.95.1
psutil==5.9.5
pydantic==1.10.2
PyJWT==2.6.0
PyJWT==2.6.0
python_bcrypt==0.3.2
python_jose==3.3.0
schema==0.7.5
starlette==0.26.1
uvicorn==0.22.0
